+++ src/components/MlSketchTool/MlSketchTool.stories.tsx
@@ -1,0 +1,26 @@
+import React from 'react';
+
+import MlSketchTool from './MlSketchTool';
+
+import mapContextDecorator from '../../decorators/MapContextDecorator';
+import Sidebar from '../../ui_components/Sidebar';
+
+const storyoptions = {
+	title: 'MapComponents/MlSketchTool',
+	component: MlSketchTool,
+	argTypes: {},
+	decorators: mapContextDecorator,
+};
+export default storyoptions;
+
+const Template = () => (
+	<>
+		<Sidebar sx={{ maxWidth: '25vw' }}>
+			<MlSketchTool />
+		</Sidebar>
+	</>
+);
+
+export const ExampleConfig = Template.bind({});
+ExampleConfig.parameters = {};
+ExampleConfig.args = {};
